+++ Magna/src/main/java/de/morigm/magna/stuff/MagnaStuff.java
@@ -1,7 +1,11 @@
 package de.morigm.magna.stuff;
 
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
+
+import org.bukkit.entity.Player;
 
 import lombok.Getter;
 
@@ -10,8 +14,10 @@ public class MagnaStuff
 	static
 	{
 		msg = new HashMap<>();
+		onlybreaks = new ArrayList<>();
 	}
 	
 	
 	@Getter private static Map<String,String> msg;
+	@Getter private static List<Player> onlybreaks;
 }